[#36] Move all localStorage to Editor to prevent re-rendering
